requirements.txt
transformers
torch
faiss-cpu
sentence-transformers
numpy
datasets
python-dotenv
openai

